refactor: update UpdateUserToModerator module with controller and service
