+++ src/domain/identity/admin/update-user-to-moderator/update-user-to-moderator.dto.ts
@@ -1,0 +1,6 @@
+import { IsUUID } from 'class-validator';
+
+export class UpdateUserToModeratorDto {
+  @IsUUID()
+    userId: string;
+}
