+++ src/domain/identity/admin/update-user-to-moderator/update-user-to-moderator.service.ts
@@ -1,13 +1,11 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
-import { Role } from '@prisma/client';
-import { UpdateUserToModeratorDto } from './update-user-to-moderator.dto';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '@/shared/prisma/prisma.service';
 
 @Injectable()
 export class UpdateUserToModeratorService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async updateUserRole({ userId, newRole }: UpdateUserToModeratorDto) {
+  async updateUserRole(userId: string) {
     const user = await this.prisma.user.findUnique({
       where: { id: userId },
     });
@@ -16,14 +14,14 @@ export class UpdateUserToModeratorService {
       throw new NotFoundException(`The user with the provided ID (${userId}) does not exist.`);
     }
 
-    if (!Object.values(Role).includes(newRole as unknown as Role)) {
-      throw new BadRequestException(`Invalid role ${newRole}`);
-    }
+    user.role.push('MODERATOR');
+
+    const newRole = user.role;
 
     return this.prisma.user.update({
       where: { id: userId },
       data: {
-        role: [newRole] as unknown as [Role],
+        role: newRole,
       },
     });
   }
